+++ src/primaite/simulator/network/hardware/nodes/host/server.py
@@ -1,4 +1,7 @@
 # © Crown-owned copyright 2025, Defence Science and Technology Laboratory UK
+
+from pydantic import Field
+
 from primaite.simulator.network.hardware.nodes.host.host_node import HostNode
 
 
@@ -30,8 +33,22 @@ class Server(HostNode, identifier="server"):
         * Web Browser
     """
 
+    class ConfigSchema(HostNode.ConfigSchema):
+        """Configuration Schema for Server class."""
+
+        hostname: str = "server"
+
+    config: ConfigSchema = Field(default_factory=lambda: Server.ConfigSchema())
+
 
 class Printer(HostNode, identifier="printer"):
     """Printer? I don't even know her!."""
 
     # TODO: Implement printer-specific behaviour
+
+    class ConfigSchema(HostNode.ConfigSchema):
+        """Configuration Schema for Printer class."""
+
+        hostname: str = "printer"
+
+    config: ConfigSchema = Field(default_factory=lambda: Printer.ConfigSchema())