+++ src/primaite/simulator/network/hardware/nodes/network/wireless_router.py
@@ -2,7 +2,7 @@
 from ipaddress import IPv4Address
 from typing import Any, Dict, Optional, Union
 
-from pydantic import validate_call
+from pydantic import Field, validate_call
 
 from primaite.simulator.network.airspace import AirSpace, AirSpaceFrequency, FREQ_WIFI_2_4, IPWirelessNetworkInterface
 from primaite.simulator.network.hardware.node_operating_state import NodeOperatingState
@@ -122,13 +122,23 @@ class WirelessRouter(Router, identifier="wireless_router"):
 
     network_interfaces: Dict[str, Union[RouterInterface, WirelessAccessPoint]] = {}
     network_interface: Dict[int, Union[RouterInterface, WirelessAccessPoint]] = {}
-    airspace: AirSpace
 
-    def __init__(self, hostname: str, airspace: AirSpace, **kwargs):
-        super().__init__(hostname=hostname, num_ports=0, airspace=airspace, **kwargs)
+    class ConfigSchema(Router.ConfigSchema):
+        """Configuration Schema for WirelessRouter nodes within PrimAITE."""
+
+        hostname: str = "WirelessRouter"
+        airspace: AirSpace
+        num_ports: int = 0
+
+    config: ConfigSchema = Field(default_factory=lambda: WirelessRouter.ConfigSchema())
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
 
         self.connect_nic(
-            WirelessAccessPoint(ip_address="127.0.0.1", subnet_mask="255.0.0.0", gateway="0.0.0.0", airspace=airspace)
+            WirelessAccessPoint(
+                ip_address="127.0.0.1", subnet_mask="255.0.0.0", gateway="0.0.0.0", airspace=kwargs["config"].airspace
+            )
         )
 
         self.connect_nic(RouterInterface(ip_address="127.0.0.1", subnet_mask="255.0.0.0", gateway="0.0.0.0"))
@@ -226,7 +236,7 @@ class WirelessRouter(Router, identifier="wireless_router"):
         )
 
     @classmethod
-    def from_config(cls, cfg: Dict, **kwargs) -> "WirelessRouter":
+    def from_config(cls, config: Dict, **kwargs) -> "WirelessRouter":
         """Generate the wireless router from config.
 
         Schema:
@@ -253,22 +263,22 @@ class WirelessRouter(Router, identifier="wireless_router"):
         :return: WirelessRouter instance.
         :rtype: WirelessRouter
         """
-        operating_state = (
-            NodeOperatingState.ON if not (p := cfg.get("operating_state")) else NodeOperatingState[p.upper()]
+        router = cls(config=cls.ConfigSchema(**config))
+        router.operating_state = (
+            NodeOperatingState.ON if not (p := config.get("operating_state")) else NodeOperatingState[p.upper()]
         )
-        router = cls(hostname=cfg["hostname"], operating_state=operating_state, airspace=kwargs["airspace"])
-        if "router_interface" in cfg:
-            ip_address = cfg["router_interface"]["ip_address"]
-            subnet_mask = cfg["router_interface"]["subnet_mask"]
+        if "router_interface" in config:
+            ip_address = config["router_interface"]["ip_address"]
+            subnet_mask = config["router_interface"]["subnet_mask"]
             router.configure_router_interface(ip_address=ip_address, subnet_mask=subnet_mask)
-        if "wireless_access_point" in cfg:
-            ip_address = cfg["wireless_access_point"]["ip_address"]
-            subnet_mask = cfg["wireless_access_point"]["subnet_mask"]
-            frequency = AirSpaceFrequency._registry[cfg["wireless_access_point"]["frequency"]]
+        if "wireless_access_point" in config:
+            ip_address = config["wireless_access_point"]["ip_address"]
+            subnet_mask = config["wireless_access_point"]["subnet_mask"]
+            frequency = AirSpaceFrequency._registry[config["wireless_access_point"]["frequency"]]
             router.configure_wireless_access_point(ip_address=ip_address, subnet_mask=subnet_mask, frequency=frequency)
 
-        if "acl" in cfg:
-            for r_num, r_cfg in cfg["acl"].items():
+        if "acl" in config:
+            for r_num, r_cfg in config["acl"].items():
                 router.acl.add_rule(
                     action=ACLAction[r_cfg["action"]],
                     src_port=None if not (p := r_cfg.get("src_port")) else PORT_LOOKUP[p],
@@ -280,8 +290,8 @@ class WirelessRouter(Router, identifier="wireless_router"):
                     dst_wildcard_mask=r_cfg.get("dst_wildcard_mask"),
                     position=r_num,
                 )
-        if "routes" in cfg:
-            for route in cfg.get("routes"):
+        if "routes" in config:
+            for route in config.get("routes"):
                 router.route_table.add_route(
                     address=IPv4Address(route.get("address")),
                     subnet_mask=IPv4Address(route.get("subnet_mask", "255.255.255.0")),