+++ src/primaite/simulator/network/networks.py
@@ -40,41 +40,45 @@ def client_server_routed() -> Network:
     network = Network()
 
     # Router 1
-    router_1 = Router(hostname="router_1", num_ports=3)
+    router_1 = Router(config=dict(hostname="router_1", num_ports=3))
     router_1.power_on()
     router_1.configure_port(port=1, ip_address="192.168.1.1", subnet_mask="255.255.255.0")
     router_1.configure_port(port=2, ip_address="192.168.2.1", subnet_mask="255.255.255.0")
 
     # Switch 1
-    switch_1 = Switch(hostname="switch_1", num_ports=6)
+    switch_1 = Switch(config=dict(hostname="switch_1", num_ports=6))
     switch_1.power_on()
     network.connect(endpoint_a=router_1.network_interface[1], endpoint_b=switch_1.network_interface[6])
     router_1.enable_port(1)
 
     # Switch 2
-    switch_2 = Switch(hostname="switch_2", num_ports=6)
+    switch_2 = Switch(config=dict(hostname="switch_2", num_ports=6))
     switch_2.power_on()
     network.connect(endpoint_a=router_1.network_interface[2], endpoint_b=switch_2.network_interface[6])
     router_1.enable_port(2)
 
     # Client 1
     client_1 = Computer(
-        hostname="client_1",
-        ip_address="192.168.2.2",
-        subnet_mask="255.255.255.0",
-        default_gateway="192.168.2.1",
-        start_up_duration=0,
+        config=dict(
+            hostname="client_1",
+            ip_address="192.168.2.2",
+            subnet_mask="255.255.255.0",
+            default_gateway="192.168.2.1",
+            start_up_duration=0,
+        )
     )
     client_1.power_on()
     network.connect(endpoint_b=client_1.network_interface[1], endpoint_a=switch_2.network_interface[1])
 
     # Server 1
     server_1 = Server(
-        hostname="server_1",
-        ip_address="192.168.1.2",
-        subnet_mask="255.255.255.0",
-        default_gateway="192.168.1.1",
-        start_up_duration=0,
+        config=dict(
+            hostname="server_1",
+            ip_address="192.168.1.2",
+            subnet_mask="255.255.255.0",
+            default_gateway="192.168.1.1",
+            start_up_duration=0,
+        )
     )
     server_1.power_on()
     network.connect(endpoint_b=server_1.network_interface[1], endpoint_a=switch_1.network_interface[1])
@@ -128,32 +132,41 @@ def arcd_uc2_network() -> Network:
     network = Network()
 
     # Router 1
-    router_1 = Router(hostname="router_1", num_ports=5, start_up_duration=0)
+    router_1 = Router.from_config(
+        config={"type": "router", "hostname": "router_1", "num_ports": 5, "start_up_duration": 0}
+    )
     router_1.power_on()
     router_1.configure_port(port=1, ip_address="192.168.1.1", subnet_mask="255.255.255.0")
     router_1.configure_port(port=2, ip_address="192.168.10.1", subnet_mask="255.255.255.0")
 
     # Switch 1
-    switch_1 = Switch(hostname="switch_1", num_ports=8, start_up_duration=0)
+    switch_1 = Switch.from_config(
+        config={"type": "switch", "hostname": "switch_1", "num_ports": 8, "start_up_duration": 0}
+    )
     switch_1.power_on()
     network.connect(endpoint_a=router_1.network_interface[1], endpoint_b=switch_1.network_interface[8])
     router_1.enable_port(1)
 
     # Switch 2
-    switch_2 = Switch(hostname="switch_2", num_ports=8, start_up_duration=0)
+    switch_2 = Switch.from_config(
+        config={"type": "switch", "hostname": "switch_2", "num_ports": 8, "start_up_duration": 0}
+    )
     switch_2.power_on()
     network.connect(endpoint_a=router_1.network_interface[2], endpoint_b=switch_2.network_interface[8])
     router_1.enable_port(2)
 
     # Client 1
-    client_1 = Computer(
-        hostname="client_1",
-        ip_address="192.168.10.21",
-        subnet_mask="255.255.255.0",
-        default_gateway="192.168.10.1",
-        dns_server=IPv4Address("192.168.1.10"),
-        start_up_duration=0,
-    )
+    client_1_cfg = {
+        "type": "computer",
+        "hostname": "client_1",
+        "ip_address": "192.168.10.21",
+        "subnet_mask": "255.255.255.0",
+        "default_gateway": "192.168.10.1",
+        "dns_server": IPv4Address("192.168.1.10"),
+        "start_up_duration": 0,
+    }
+    client_1: Computer = Computer.from_config(config=client_1_cfg)
+
     client_1.power_on()
     network.connect(endpoint_b=client_1.network_interface[1], endpoint_a=switch_2.network_interface[1])
     client_1.software_manager.install(DatabaseClient)
@@ -172,14 +185,18 @@ def arcd_uc2_network() -> Network:
     )
 
     # Client 2
-    client_2 = Computer(
-        hostname="client_2",
-        ip_address="192.168.10.22",
-        subnet_mask="255.255.255.0",
-        default_gateway="192.168.10.1",
-        dns_server=IPv4Address("192.168.1.10"),
-        start_up_duration=0,
-    )
+
+    client_2_cfg = {
+        "type": "computer",
+        "hostname": "client_2",
+        "ip_address": "192.168.10.22",
+        "subnet_mask": "255.255.255.0",
+        "default_gateway": "192.168.10.1",
+        "dns_server": IPv4Address("192.168.1.10"),
+        "start_up_duration": 0,
+    }
+    client_2: Computer = Computer.from_config(config=client_2_cfg)
+
     client_2.power_on()
     client_2.software_manager.install(DatabaseClient)
     db_client_2 = client_2.software_manager.software.get("DatabaseClient")
@@ -193,27 +210,36 @@ def arcd_uc2_network() -> Network:
     )
 
     # Domain Controller
-    domain_controller = Server(
-        hostname="domain_controller",
-        ip_address="192.168.1.10",
-        subnet_mask="255.255.255.0",
-        default_gateway="192.168.1.1",
-        start_up_duration=0,
-    )
+
+    domain_controller_cfg = {
+        "type": "server",
+        "hostname": "domain_controller",
+        "ip_address": "192.168.1.10",
+        "subnet_mask": "255.255.255.0",
+        "default_gateway": "192.168.1.1",
+        "start_up_duration": 0,
+    }
+
+    domain_controller = Server.from_config(config=domain_controller_cfg)
     domain_controller.power_on()
     domain_controller.software_manager.install(DNSServer)
 
     network.connect(endpoint_b=domain_controller.network_interface[1], endpoint_a=switch_1.network_interface[1])
 
     # Database Server
-    database_server = Server(
-        hostname="database_server",
-        ip_address="192.168.1.14",
-        subnet_mask="255.255.255.0",
-        default_gateway="192.168.1.1",
-        dns_server=IPv4Address("192.168.1.10"),
-        start_up_duration=0,
-    )
+
+    database_server_cfg = {
+        "type": "server",
+        "hostname": "database_server",
+        "ip_address": "192.168.1.14",
+        "subnet_mask": "255.255.255.0",
+        "default_gateway": "192.168.1.1",
+        "dns_server": IPv4Address("192.168.1.10"),
+        "start_up_duration": 0,
+    }
+
+    database_server = Server.from_config(config=database_server_cfg)
+
     database_server.power_on()
     network.connect(endpoint_b=database_server.network_interface[1], endpoint_a=switch_1.network_interface[3])
 
@@ -223,14 +249,18 @@ def arcd_uc2_network() -> Network:
     database_service.configure_backup(backup_server=IPv4Address("192.168.1.16"))
 
     # Web Server
-    web_server = Server(
-        hostname="web_server",
-        ip_address="192.168.1.12",
-        subnet_mask="255.255.255.0",
-        default_gateway="192.168.1.1",
-        dns_server=IPv4Address("192.168.1.10"),
-        start_up_duration=0,
-    )
+
+    web_server_cfg = {
+        "type": "server",
+        "hostname": "web_server",
+        "ip_address": "192.168.1.11",
+        "subnet_mask": "255.255.255.0",
+        "default_gateway": "192.168.1.1",
+        "dns_server": IPv4Address("192.168.1.10"),
+        "start_up_duration": 0,
+    }
+    web_server = Server.from_config(config=web_server_cfg)
+
     web_server.power_on()
     web_server.software_manager.install(DatabaseClient)
 
@@ -247,27 +277,32 @@ def arcd_uc2_network() -> Network:
     dns_server_service.dns_register("arcd.com", web_server.network_interface[1].ip_address)
 
     # Backup Server
-    backup_server = Server(
-        hostname="backup_server",
-        ip_address="192.168.1.16",
-        subnet_mask="255.255.255.0",
-        default_gateway="192.168.1.1",
-        dns_server=IPv4Address("192.168.1.10"),
-        start_up_duration=0,
-    )
+    backup_server_cfg = {
+        "type": "server",
+        "hostname": "backup_server",
+        "ip_address": "192.168.1.16",
+        "subnet_mask": "255.255.255.0",
+        "default_gateway": "192.168.1.1",
+        "dns_server": IPv4Address("192.168.1.10"),
+        "start_up_duration": 0,
+    }
+    backup_server: Server = Server.from_config(config=backup_server_cfg)
+
     backup_server.power_on()
     backup_server.software_manager.install(FTPServer)
     network.connect(endpoint_b=backup_server.network_interface[1], endpoint_a=switch_1.network_interface[4])
 
     # Security Suite
-    security_suite = Server(
-        hostname="security_suite",
-        ip_address="192.168.1.110",
-        subnet_mask="255.255.255.0",
-        default_gateway="192.168.1.1",
-        dns_server=IPv4Address("192.168.1.10"),
-        start_up_duration=0,
-    )
+    security_suite_cfg = {
+        "type": "server",
+        "hostname": "backup_server",
+        "ip_address": "192.168.1.110",
+        "subnet_mask": "255.255.255.0",
+        "default_gateway": "192.168.1.1",
+        "dns_server": IPv4Address("192.168.1.10"),
+        "start_up_duration": 0,
+    }
+    security_suite: Server = Server.from_config(config=security_suite_cfg)
     security_suite.power_on()
     network.connect(endpoint_b=security_suite.network_interface[1], endpoint_a=switch_1.network_interface[7])
     security_suite.connect_nic(NIC(ip_address="192.168.10.110", subnet_mask="255.255.255.0"))