+++ tests/integration_tests/game_layer/observations/test_node_observations.py
@@ -25,7 +25,7 @@ def test_host_observation(simulation):
     pc: Computer = simulation.network.get_node_by_hostname("client_1")
 
     host_obs = HostObservation(
-        where=["network", "nodes", pc.hostname],
+        where=["network", "nodes", pc.config.hostname],
         num_applications=0,
         num_files=1,
         num_folders=1,
@@ -56,7 +56,7 @@ def test_host_observation(simulation):
     observation_state = host_obs.observe(simulation.describe_state())
     assert observation_state.get("operating_status") == 4  # shutting down
 
-    for i in range(pc.shut_down_duration + 1):
+    for i in range(pc.config.shut_down_duration + 1):
         pc.apply_timestep(i)
 
     observation_state = host_obs.observe(simulation.describe_state())