+++ tests/integration_tests/test_simulation/test_request_response.py
@@ -111,7 +111,7 @@ def test_request_fails_if_node_off(example_network, node_request):
     """Test that requests succeed when the node is on, and fail if the node is off."""
     net = example_network
     client_1: HostNode = net.get_node_by_hostname("client_1")
-    client_1.shut_down_duration = 0
+    client_1.config.shut_down_duration = 0
 
     assert client_1.operating_state == NodeOperatingState.ON
     resp_1 = net.apply_request(node_request)
@@ -140,9 +140,9 @@ class TestDataManipulationGreenRequests:
         client_1 = net.get_node_by_hostname("client_1")
         client_2 = net.get_node_by_hostname("client_2")
 
-        client_1.shut_down_duration = 0
+        client_1.config.shut_down_duration = 0
         client_1.power_off()
-        client_2.shut_down_duration = 0
+        client_2.config.shut_down_duration = 0
         client_2.power_off()
 
         client_1_browser_execute_off = net.apply_request(["node", "client_1", "application", "WebBrowser", "execute"])