+++ tests/unit_tests/_primaite/_simulator/_system/_services/test_dns_client.py
@@ -14,13 +14,15 @@ from primaite.utils.validation.port import PORT_LOOKUP
 
 @pytest.fixture(scope="function")
 def dns_client() -> Computer:
-    node = Computer(
-        hostname="dns_client",
-        ip_address="192.168.1.11",
-        subnet_mask="255.255.255.0",
-        default_gateway="192.168.1.1",
-        dns_server=IPv4Address("192.168.1.10"),
-    )
+    node_cfg = {
+        "type": "computer",
+        "hostname": "dns_client",
+        "ip_address": "192.168.1.11",
+        "subnet_mask": "255.255.255.0",
+        "default_gateway": "192.168.1.1",
+        "dns_server": IPv4Address("192.168.1.10"),
+    }
+    node = Computer.from_config(config=node_cfg)
     return node
 
 
@@ -34,6 +36,16 @@ def test_create_dns_client(dns_client):
 
 def test_dns_client_add_domain_to_cache_when_not_running(dns_client):
     dns_client_service: DNSClient = dns_client.software_manager.software.get("DNSClient")
+
+    # shutdown the dns_client
+    dns_client.power_off()
+
+    # wait for dns_client to turn off
+    idx = 0
+    while dns_client.operating_state == NodeOperatingState.SHUTTING_DOWN:
+        dns_client.apply_timestep(idx)
+        idx += 1
+
     assert dns_client.operating_state is NodeOperatingState.OFF
     assert dns_client_service.operating_state is ServiceOperatingState.STOPPED
 
@@ -61,7 +73,7 @@ def test_dns_client_check_domain_exists_when_not_running(dns_client):
 
     dns_client.power_off()
 
-    for i in range(dns_client.shut_down_duration + 1):
+    for i in range(dns_client.config.shut_down_duration + 1):
         dns_client.apply_timestep(timestep=i)
 
     assert dns_client.operating_state is NodeOperatingState.OFF